Criação do Delete 1 item dentro de um Carrinho, e feito a manuntenção da função create de CartItemService, onde ele valida se um caarrinho está vazio, ou não, e assim altera ou cria um item dentro do carrinho conforme regra de negócio.

+++ src/main/java/com/example/ecommerce/modules/cart/dto/CartItemAddProductDTO.java
@@ -2,6 +2,6 @@ package com.example.ecommerce.modules.cart.dto;
 
 
 public record CartItemAddProductDTO(
-        Long productId
+        Integer productId
 ) {
 }
